updated jdk version and added webview backpress support
The JDK version was updated from 17 to 21 in the project's configuration files.
Additionally, the `MainActivity` was modified to handle the back button press within the `WebView`, allowing navigation to the previous page if available.
The Javascript interface name was also changed from "Android" to "Fahim".
Gradle project settings were updated to set the test runner to "CHOOSE_PER_TEST" and `resolveExternalAnnotations` was removed.