quiz feature about the father of java
The user can select an answer from the radio buttons and submit it.
The app will then display whether the answer is correct or not.
The correct answer is "James Gosling".
If the user selects the correct answer, the text color will be green.
Otherwise, the text color will be red.